+++ app/javascript/mastodon/features/account_timeline/components/tabs.tsx
@@ -1,0 +1,27 @@
+import type { FC } from 'react';
+
+import { FormattedMessage } from 'react-intl';
+
+import { NavLink } from 'react-router-dom';
+
+export const AccountTabs: FC<{ acct: string }> = ({ acct }) => {
+  return (
+    <div className='account__section-headline'>
+      <NavLink exact to={`/@${acct}/featured`}>
+        <FormattedMessage id='account.featured' defaultMessage='Featured' />
+      </NavLink>
+      <NavLink exact to={`/@${acct}`}>
+        <FormattedMessage id='account.posts' defaultMessage='Posts' />
+      </NavLink>
+      <NavLink exact to={`/@${acct}/with_replies`}>
+        <FormattedMessage
+          id='account.posts_with_replies'
+          defaultMessage='Posts and replies'
+        />
+      </NavLink>
+      <NavLink exact to={`/@${acct}/media`}>
+        <FormattedMessage id='account.media' defaultMessage='Media' />
+      </NavLink>
+    </div>
+  );
+};
